+++ contaDuracao.h
@@ -1,8 +1,8 @@
 #include <vector>
 using namespace std;
 
-int getTime(vector<int> arr, int metodo, int& cont);
-int getTimeMerge(vector<int> arr, int& cont);
-int getTimeShell(vector<int> arr, int& cont);
-int getTimeInsertion(vector<int> arr, int& cont);
-int getTimeBubble(vector<int> arr, int& cont);
+int getTime(vector<int> arr, int metodo, long long& cont);
+int getTimeMerge(vector<int> arr, long long& cont);
+int getTimeShell(vector<int> arr, long long& cont);
+int getTimeInsertion(vector<int> arr, long long& cont);
+int getTimeBubble(vector<int> arr, long long& cont);
